% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_plot.R
\name{Gibbs_plot}
\alias{Gibbs_plot}
\title{Gibbs Plot Function}
\usage{
Gibbs_plot(N, burnin, sigma, mu, set_seed = 2012)
}
\arguments{
\item{N}{An integer for the number of iterations.}

\item{burnin}{An integer for the number of burn-in iterations.}

\item{sigma}{A covariance matrix.}

\item{mu}{A numeric vector for the mean values.}

\item{set_seed}{An optional integer for the RNG seed.}
}
\value{
A list containing a linear model object and a Shapiro-Wilk test object.
}
\description{
This function performs Gibbs sampling and creates plots for the sampling chain.
}
\examples{
\dontrun{
results <- Gibbs_plot(N = 2000, burnin = 500, sigma = matrix(c(1,0.9,0.9,1),
                      nrow=2), mu = c(0,0))
# View summary information for linear models
summary(results$fit)

# View the results of the Shapiro-Wilk normality test
results$shapiro
}
}
