% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gibbs_sampling_cpp}
\alias{Gibbs_sampling_cpp}
\title{A Gibbs sampler using Rcpp}
\usage{
Gibbs_sampling_cpp(N, burnin, sigma, mu)
}
\arguments{
\item{N}{Total number of samples to generate}

\item{burnin}{Number of initial samples to discard}

\item{sigma}{Covariance matrix for the bivariate normal distribution}

\item{mu}{Mean vector for the bivariate normal distribution}
}
\value{
A numeric matrix with each row being a sample from the distribution
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
# Define the mean vector and covariance matrix for a bivariate normal distribution
mu <- c(0, 0)
sigma <- matrix(c(1, 0.5, 0.5, 1), nrow = 2, byrow = TRUE)
# Run Gibbs sampling to generate samples
set.seed(123) # Set seed for reproducibility
samples <- Gibbs_sampling_cpp(N = 1000, burnin = 100, sigma = sigma, mu = mu)
}
}
