\name{SA23229024-package}
\alias{SA23229024-package}
\docType{package}
\title{This package provides Several R and Rcpp function for the 'Statistical Computing' }
\description{
Several R and Rcpp function are included in this package for the 'Statistical Computing' course in USTC
}
\author{
Liu jingtao
Maintainer: Liu jingtao <jingt-liu@mail.ustc.edu.cn>
}
\keyword{ package }
